add inference script & overiding args for llama

+++ src/models/layers/blockdiag_butterfly_multiply.py
@@ -19,7 +19,7 @@ class BlockdiagMultiply(torch.autograd.Function):
     """
 
     @staticmethod
-    @torch.cuda.amp.custom_fwd(cast_inputs=torch.bfloat16)
+    @torch.cuda.amp.custom_fwd(cast_inputs=torch.float16)
     def forward(ctx, x, weight):
         ctx.save_for_backward(x, weight)
         batch_shape, n = x.shape[:-1], x.shape[-1]
